pyproject.toml
[tool.poetry]
name = "myapp"
version = "0.1.0"
description = "My application built with MoWaKi 2.0"
authors = ["Samuele Santi <samuele@samuelesanti.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {extras = ["asgi", "debug-server"], version = "^0.215.1"}
starlette = "^0.32.0.post1"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"
pytz = "^2023.3.post1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.29.0"
alembic = "^1.12.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.3"
ipython = "^8.18.1"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
